backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pandas==2.1.4
requests==2.31.0
email-validator==2.1.0
pymongo==4.6.0
redis==5.0.1
psycopg2-binary==2.9.9
